@@ -21,7 +21,7 @@
     "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.14.19",
     "axios": "^1.6.2",
-    "next": "13.4.19",
+    "next": "^14.0.5-canary.1",
     "react": "^18",
     "react-dom": "^18",
     "react-hot-toast": "^2.4.1",
